# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tabaq | طبق - Food & Dining Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: restaurantId
          in: query
          schema:
            type: integer
        - name: dishId
          in: query
          schema:
            type: integer
        - name: userId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewRequest"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{reviewId}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get review details
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /reviews/{reviewId}/like:
    post:
      operationId: likeReview
      tags: [reviews]
      summary: Like a review
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeStatusResponse"

  # ─── Users ────────────────────────────────────────────────────────────────
...
  /search:
    get:
      operationId: search
      tags: [search]
      summary: Search restaurants, dishes, and venues
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [all, restaurants, dishes, venues]
            default: all
        - name: cityId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
  /search/autocomplete:
    get:
      operationId: autocomplete
      tags: [search]
      summary: Autocomplete search suggestions
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: cityId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Autocomplete suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutocompleteResponse"

  # ─── Events ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    RestaurantCard:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        coverImageUrl:
          type: string
        priceTier:
          type: string
          enum: [budget, mid, upscale, fine_dining]
        avgRating:
          type: number
        reviewCount:
          type: integer
        isVerified:
          type: boolean
        cityId:
          type: integer
        cuisineTypes:
          type: array
          items:
            type: string
      required:
        - id
        - nameEn
        - nameAr
        - priceTier
        - avgRating
        - reviewCount
        - isVerified
        - cityId
        - cuisineTypes
...
    DishCard:
      type: object
      properties:
        id:
          type: integer
        nameEn:
          type: string
        nameAr:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        currency:
          type: string
        avgRating:
          type: number
        reviewCount:
          type: integer
        restaurantNameEn:
          type: string
        restaurantNameAr:
          type: string
        restaurantId:
          type: integer
        cityId:
          type: integer
      required:
        - id
        - nameEn
        - nameAr
        - avgRating
        - reviewCount
        - restaurantNameEn
        - restaurantNameAr
        - restaurantId
        - cityId
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userNameEn:
          type: string
        userNameAr:
          type: string
        userAvatarUrl:
          type: string
        userLevel:
          type: integer
        userLevelTitle:
          type: string
        restaurantId:
          type: integer
        dishId:
          type: integer
        ratingOverall:
          type: number
          minimum: 1
          maximum: 5
        ratingFood:
          type: number
          minimum: 1
          maximum: 5
        ratingService:
          type: number
          minimum: 1
          maximum: 5
        ratingAmbiance:
          type: number
          minimum: 1
          maximum: 5
        ratingValue:
          type: number
          minimum: 1
          maximum: 5
        textEn:
          type: string
        textAr:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        likeCount:
          type: integer
        isLiked:
          type: boolean
        visitDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userNameEn
        - ratingOverall
        - likeCount
        - isLiked
        - createdAt
    CreateReviewRequest:
      type: object
      properties:
        restaurantId:
          type: integer
        dishId:
          type: integer
        ratingOverall:
          type: number
          minimum: 1
          maximum: 5
        ratingFood:
          type: number
          minimum: 1
          maximum: 5
        ratingService:
          type: number
          minimum: 1
          maximum: 5
        ratingAmbiance:
          type: number
          minimum: 1
          maximum: 5
        ratingValue:
          type: number
          minimum: 1
          maximum: 5
        textEn:
          type: string
        textAr:
          type: string
        photoUrls:
          type: array
          items:
            type: string
        visitDate:
          type: string
          format: date
      required:
        - ratingOverall
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
      required:
        - reviews
        - total
        - offset
        - limit
    LikeStatusResponse:
      type: object
      properties:
        isLiked:
          type: boolean
        likeCount:
          type: integer
      required:
        - isLiked
        - likeCount

    # ─── Search ───────────────────────────────────────────────────────────
    SearchResponse:
      type: object
      properties:
        restaurants:
          type: array
          items:
            $ref: "#/components/schemas/RestaurantCard"
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/DishCard"
        totalRestaurants:
          type: integer
        totalDishes:
          type: integer
      required:
        - restaurants
        - dishes
        - totalRestaurants
        - totalDishes
    AutocompleteResponse:
      type: object
      properties:
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/AutocompleteSuggestion"
      required:
        - suggestions
    AutocompleteSuggestion:
      type: object
      properties:
        type:
          type: string
          enum: [restaurant, dish, city, category]
        id:
          type: integer
        labelEn:
          type: string
        labelAr:
          type: string
        imageUrl:
          type: string
      required:
        - type
        - id
        - labelEn
        - labelAr

    # ─── Events ───────────────────────────────────────────────────────────
...